requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
chromadb>=0.4.0
sentence-transformers==2.2.2
PyPDF2>=3.0.0
python-docx>=0.8.11
pandas>=1.5.0
numpy>=1.24.0,<2.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
torch>=2.2.0
transformers>=4.21.0,<5.0.0
tokenizers>=0.13.0
huggingface_hub>=0.15.0,<0.20.0
scikit-learn>=1.3.0
